+++ src/types/dbus.ts
@@ -1,7 +1,7 @@
 import Gio from "gi://Gio?version=2.0";
 import GLib from "gi://GLib?version=2.0";
-import { KeysOf } from "./common.js";
-import { LoopStatus, PlaybackStatus } from "./enums/panel.js";
+import { KeysOf } from "./general.js";
+import { LoopStatus, PlaybackStatus } from "./enums/shell.js";
 
 type MethodResult<T> = [T];
